refactor: Break up db module into submodules for better clarity
