+++ src/db/linkers/__init__.py
@@ -1,0 +1,6 @@
+from src.db.linkers.paper_dataset_linker import PaperDatasetLinker
+from src.db.linkers.europepmc_dataset_linker import EuropePMCDatasetLinker
+from src.db.linkers.elink_dataset_linker import ELinkDatasetLinker
+from src.db.linkers.chained_dataset_linker import ChainedDatasetLinker
+
+__all__ = ["PaperDatasetLinker", "EuropePMCDatasetLinker", "ELinkDatasetLinker", "ChainedDatasetLinker"]
