+++ src/db/loaders/ncbi_gse_loader.py
@@ -5,10 +5,10 @@ import GEOparse
 import requests
 from dacite import from_dict
 
-from src.db.geoparse_to_geometadb import format_geoparse_metadata
-from src.db.gse import GSE
-from src.db.gse_loader import GSELoader
-from src.db.gse_repository import GSERepository
+from src.db.loaders.geoparse_to_geometadb import format_geoparse_metadata
+from src.db.models.gse import GSE
+from src.db.loaders.gse_loader import GSELoader
+from src.db.repositories.gse_repository import GSERepository
 from src.exception.geo_error import GEOError
 
 logger = logging.getLogger(__name__)
