+++ src/db/models/__init__.py
@@ -1,0 +1,6 @@
+from src.db.models.gse import GSE, GSE_DTO
+from src.db.models.gsm import GSM
+from src.db.models.gse_gsm import GSE_GSM
+from src.db.models.mapper_registry import mapper_registry
+
+__all__ = ["GSE", "GSE_DTO", "GSM", "GSE_GSM", "mapper_registry"]
