+++ src/db/repositories/__init__.py
@@ -1,0 +1,3 @@
+from src.db.repositories.gse_repository import GSERepository
+
+__all__ = ["GSERepository"]
