+++ src/test/db/loaders/test_ncbi_gse_loader.py
@@ -1,13 +1,13 @@
 import unittest
 from typing import List
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import requests
 from parameterized import parameterized
 
 from src.config.config import Config
-from src.db.gse import GSE
-from src.db.ncbi_gse_loader import NCBIGSELoader
+from src.db.loaders.ncbi_gse_loader import NCBIGSELoader
+from src.db.models.gse import GSE
 from src.exception.geo_error import GEOError
 from src.test.helpers.http import create_mock_response
